+++ mobile/apps/locker/lib/ui/drawer/drawer_page.dart
@@ -1,12 +1,13 @@
 import "package:ente_components/ente_components.dart";
 import "package:ente_sharing/verify_identity_dialog.dart";
 import "package:ente_strings/ente_strings.dart";
+import "package:ente_ui/pages/settings_search_page.dart";
 import "package:flutter/material.dart";
 import "package:hugeicons/hugeicons.dart";
 import "package:locker/services/configuration.dart";
 import "package:locker/ui/components/legacy_collections_trash_widget.dart";
 import "package:locker/ui/components/usage_card_widget.dart";
-import "package:locker/ui/settings/pages/settings_search_page.dart";
+import "package:locker/ui/settings/search/settings_search_registry.dart";
 import "package:locker/ui/settings/settings_page.dart";
 
 class DrawerPage extends StatelessWidget {
@@ -92,9 +93,20 @@ class DrawerPage extends StatelessWidget {
   }
 
   void _openSearch(BuildContext context) {
-    Navigator.of(
-      context,
-    ).push(MaterialPageRoute(builder: (context) => const SettingsSearchPage()));
+    Navigator.of(context).push(
+      MaterialPageRoute(
+        builder: (context) {
+          final items = SettingsSearchRegistry.getSearchableItems(context);
+          return SettingsSearchPage(
+            items: items,
+            suggestions: SettingsSearchRegistry.getSuggestions(context, items),
+            onNavigate: (context, routeBuilder) => Navigator.of(
+              context,
+            ).push(MaterialPageRoute(builder: routeBuilder)),
+          );
+        },
+      ),
+    );
   }
 
   void _showVerifyIDSheet(BuildContext context) {
