+++ mobile/apps/photos/lib/ui/settings/search/settings_search_registry.dart
@@ -1,6 +1,7 @@
 import "dart:io";
 
 import "package:ente_strings/ente_strings.dart";
+import "package:ente_ui/models/settings_search_item.dart";
 import "package:flutter/foundation.dart";
 import "package:flutter/material.dart";
 import "package:hugeicons/hugeicons.dart";
@@ -15,7 +16,6 @@ import "package:photos/ui/settings/gallery_settings_screen.dart";
 import "package:photos/ui/settings/memories_settings_screen.dart";
 import "package:photos/ui/settings/ml/machine_learning_settings_page.dart";
 import "package:photos/ui/settings/notification_settings_screen.dart";
-import "package:photos/ui/settings/search/settings_search_item.dart";
 import "package:photos/ui/settings/security/security_settings_page.dart";
 import "package:photos/ui/settings/streaming/video_streaming_settings_page.dart";
 import "package:photos/ui/settings/support/help_support_page.dart";
@@ -589,6 +589,7 @@ class SettingsSearchRegistry {
           icon: HugeIcons.strokeRoundedRocket01,
           routeBuilder: (_) => const FreeUpSpaceOptionsScreen(),
           isSubPage: true,
+          sectionItemPriority: 1,
           keywords: ["device space", "storage", "delete"],
         ),
         SettingsSearchItem(
@@ -633,6 +634,7 @@ class SettingsSearchRegistry {
           icon: HugeIcons.strokeRoundedDelete02,
           routeBuilder: (_) => const FreeUpSpaceOptionsScreen(),
           isSubPage: true,
+          sectionItemPriority: 0,
           keywords: ["delete", "suggestions", "cleanup"],
         ),
         SettingsSearchItem(
@@ -770,10 +772,7 @@ class SettingsSearchRegistry {
   }
 
   /// Get suggestions shown when search is empty
-  static List<SettingsSearchSuggestion> getSuggestions(
-    BuildContext context,
-    void Function(Widget Function(BuildContext) routeBuilder) onNavigate,
-  ) {
+  static List<SettingsSearchSuggestion> getSuggestions(BuildContext context) {
     final l10n = context.strings;
     final hasLoggedIn = Configuration.instance.isLoggedIn();
     final isLocalGallery = isLocalGalleryMode;
@@ -782,27 +781,25 @@ class SettingsSearchRegistry {
       // Gallery suggestion
       SettingsSearchSuggestion(
         title: l10n.gallery,
-        onTap: () => onNavigate(
-          (_) =>
-              const GallerySettingsScreen(fromGalleryLayoutSettingsCTA: false),
-        ),
+        routeBuilder: (_) =>
+            const GallerySettingsScreen(fromGalleryLayoutSettingsCTA: false),
       ),
       // App lock suggestion
       SettingsSearchSuggestion(
         title: l10n.appLock,
-        onTap: () => onNavigate((_) => const SecuritySettingsPage()),
+        routeBuilder: (_) => const SecuritySettingsPage(),
       ),
       // Free up device space suggestion
       if (hasLoggedIn && !isLocalGallery)
         SettingsSearchSuggestion(
           title: l10n.freeUpDeviceSpace,
-          onTap: () => onNavigate((_) => const FreeUpSpaceOptionsScreen()),
+          routeBuilder: (_) => const FreeUpSpaceOptionsScreen(),
         ),
       // Backup settings suggestion
       if (hasLoggedIn && !isLocalGallery)
         SettingsSearchSuggestion(
           title: l10n.backupSettings,
-          onTap: () => onNavigate((_) => const BackupSettingsPage()),
+          routeBuilder: (_) => const BackupSettingsPage(),
         ),
     ];
   }
